third_party/WebKit/LayoutTests/http/tests/serviceworker/registration-expected.txt

Tests if ServiceWorker registration is working

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Registering normal pattern
PASS 'DisabledError' is regError.name
Registering pattern outside domain
PASS 'SecurityError' is regError.name
Registering pattern outside domain
PASS 'SecurityError' is regError.name
PASS successfullyParsed is true

TEST COMPLETE
